% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banana-package.R
\docType{package}
\name{banana-package}
\alias{banana}
\alias{banana-package}
\title{banana: Round Numeric Values to the Nearest Integer}
\description{
In R the default choice for rounding follows the IEC 60559 standard where a 5 is expected to be rounded to the even digit. This package offers alternative functionality to round to the closest integer.
}
\keyword{internal}
